Remove bundle hooks (#1658)
