+++ packages/liveblocks-react/src/suspense.ts
@@ -13,16 +13,13 @@ export { shallow } from "@liveblocks/client";
 
 // Export all the top-level hooks
 export {
+  ClientContext,
   createLiveblocksContext,
   LiveblocksProvider,
   useClient,
-  useLiveblocksContextBundle,
   useInboxNotificationThread,
   useMarkAllInboxNotificationsAsRead,
   useMarkInboxNotificationAsRead,
-
-  // TODO Move these private APIs into another namespace eventually
-  useLiveblocksContextBundleOrNull as useLiveblocksContextBundleOrNull__,
 } from "./liveblocks";
 export {
   createRoomContext,
@@ -49,7 +46,6 @@ export {
   useRedo,
   useRemoveReaction,
   useRoom,
-  useRoomContextBundle,
   useRoomNotificationSettings,
   useStatus,
   useStorageRoot,
@@ -57,9 +53,6 @@ export {
   useUndo,
   useUpdateMyPresence,
   useUpdateRoomNotificationSettings,
-
-  // TODO Move these private APIs into another namespace eventually
-  useRoomContextBundleOrNull as useRoomContextBundleOrNull__,
 } from "./room";
 
 // Export the Suspense versions of our hooks
